+++ spec/azk_benchmark_spec.js
@@ -3,19 +3,19 @@ import AzkBenchmark from '../src/benchmark/azk_benchmark';
 
 describe('AzkBenchmark:', () => {
 
-  it('should azk_bin_path have a value', () => {
+  it('should azkBinPath have a value', () => {
     let instantiate = () => {
-      new AzkBenchmark({ azk_bin_path: 'something not null' });
+      new AzkBenchmark({ azkBinPath: 'something not null' });
     };
     return h.expect(instantiate).to.not.throw(Error);
   });
 
-  it('should azk_bin_path throw if undefined/null', () => {
+  it('should azkBinPath throw if undefined/null', () => {
     let instantiate = () => {
-      new AzkBenchmark({ azk_bin_path: null });
+      new AzkBenchmark({ azkBinPath: null });
     };
     return h.expect(instantiate).to
-      .throw(Error, /azk_bin_path \(.*?\) cannot be null\/undefined/);
+      .throw(Error, /azkBinPath \(.*?\) cannot be null\/undefined/);
   });
 
 });