+++ src/benchmark/azk_benchmark.js
@@ -19,23 +19,23 @@ export default class AzkBenchmark {
     this.sendData = new SendData(this.opts);
 
     // load actions
-    this.pre_actions = actions.pre_actions;
-    this.pre_actions_prefix = (
+    this.preActions = actions.preActions;
+    this.preActionsPrefix = (
       chalk.gray('[') +
       chalk.blue('provision') +
       chalk.gray(']') +
       chalk.white(' $> ')
     );
 
-    this.main_actions = actions.main_actions;
-    this.main_actions_prefix = (
+    this.mainActions = actions.mainActions;
+    this.mainActionsPrefix = (
       chalk.gray('[') +
       chalk.blue('benchmarking') +
       chalk.gray(']') +
       chalk.white(' $> ')
     );
 
-    this.get_version_prefix = (
+    this.getVersionPrefix = (
       chalk.gray('[') +
       chalk.blue('azk version') +
       chalk.gray(']') +
@@ -44,8 +44,8 @@ export default class AzkBenchmark {
   }
 
   _validade(opts) {
-    if (!opts.azk_bin_path) {
-      throw new Error(`azk_bin_path (${opts.azk_bin_path}) cannot be null/undefined`);
+    if (!opts.azkBinPath) {
+      throw new Error(`azkBinPath (${opts.azkBinPath}) cannot be null/undefined`);
     }
   }
 
@@ -55,31 +55,30 @@ export default class AzkBenchmark {
     .then(this._runMainActions.bind(this))
     .then(this._processResults.bind(this))
     .catch((err) => {
-      const newError = new Error(err.message);
-      newError.code = err.code;
-      throw newError;
+      err.code = err.code;
+      throw err;
     });
   }
 
   _runPreActions() {
-    return BB.Promise.mapSeries(this.pre_actions, (params) => {
+    return BB.Promise.mapSeries(this.preActions, (params) => {
       let start = this._startTimer();
       let params_result = params(this.opts);
       return this._spawnCommand(
         params_result,
-        this.pre_actions_prefix,
-        this.opts.verbose_level
+        this.preActionsPrefix,
+        this.opts.verboseLevel
       )
       .then((result) => {
-        let result_to_send = {
+        let resultToSend = {
           command: 'azk ' + params_result.join(' '),
           result: result,
           time: this._stopTimer(start)
         };
-        if (this.opts.verbose_level > 0) {
-          process.stdout.write(' ' + chalk.green(result_to_send.time.toString() + 'ms') + '\n\n');
+        if (this.opts.verboseLevel > 0) {
+          process.stdout.write(' ' + chalk.green(resultToSend.time.toString() + 'ms') + '\n\n');
         }
-        return result_to_send;
+        return resultToSend;
       });
     });
   }
@@ -95,43 +94,59 @@ export default class AzkBenchmark {
     return time;
   }
 
-  _spawnCommand(params, prefix, verbose_level) {
+  _spawnCommand(params, prefix, verboseLevel) {
     return spawnAsync({
-      cwd          : this.opts.dest_path,
-      executable   : this.opts.azk_bin_path,
-      params_array : params,
+      cwd          : this.opts.destPath,
+      executable   : this.opts.azkBinPath,
+      paramsArray : params,
       prefix       : prefix,
-      verbose_level: verbose_level - 1,
+      verboseLevel: verboseLevel - 1,
     });
   }
 
   _getAzkVersion() {
-    return this._spawnCommand(['version'], this.get_version_prefix, this.opts.verbose_level - 1)
-    .then((version_result) => {
-      this.azk_version = matchFirstRegex(version_result.message, /(\d+\.\d+\.\d+)/)[1];
+    return this._spawnCommand(['version'], this.getVersionPrefix, this.opts.verboseLevel - 1)
+    .then((versionResult) => {
+      this.azk_version = matchFirstRegex(versionResult.message, /(\d+\.\d+\.\d+)/)[1];
     });
   }
 
-  _runMainActions() {
-    return BB.Promise.mapSeries(this.main_actions, (params) => {
+  _getMainActionsPromise() {
+    return BB.Promise.mapSeries(this.mainActions, (params) => {
       let start = this._startTimer();
-      return this._spawnCommand(params, this.main_actions_prefix, this.opts.verbose_level)
+      return this._spawnCommand(params, this.mainActionsPrefix, this.opts.verboseLevel)
       .then((result) => {
-        let result_to_send = {
+        let resultToSend = {
           command: 'azk ' + params.join(' '),
           result: result,
           azk_version: this.azk_version,
           time: this._stopTimer(start)
         };
-        if (this.opts.verbose_level > 0) {
-          process.stdout.write(' ' + chalk.green(result_to_send.time.toString() + 'ms') + '\n\n');
+        if (this.opts.verboseLevel > 0) {
+          process.stdout.write(' ' + chalk.green(resultToSend.time.toString() + 'ms') + '\n\n');
         }
-        return result_to_send;
+        return resultToSend;
       });
     });
   }
 
-  _processResults(final_results) {
+  _runMainActions() {
+    let runTimes = 1;
+    let mainActionsList = [];
+    if (this.opts.repeat > 1) {
+      runTimes = this.opts.repeat;
+    }
+
+    for (let i = 0; i < runTimes; i++) {
+      mainActionsList.push(this._getMainActionsPromise.bind(this));
+    }
+
+    return BB.Promise.mapSeries(mainActionsList, (mainPromise) => {
+      return mainPromise();
+    });
+  }
+
+  _processResults(finalResultsList) {
     let table_args = (this.opts.plain) ? {
       chars: { 'top': '' , 'top-mid': '' , 'top-left': '' , 'top-right': '',
        'bottom': '' , 'bottom-mid': '' , 'bottom-left': '' , 'bottom-right': '',
@@ -144,48 +159,61 @@ export default class AzkBenchmark {
 
     var table = new Table(table_args);
 
-    // each result
-    final_results.forEach((item) => {
-      let row = {};
-      row[item.command] = item.time;
-      table.push(row);
+    // add each
+    finalResultsList.forEach((finalResults) => {
+      // each result
+      finalResults.forEach((item) => {
+        let row = {};
+        row[item.command] = item.time;
+        table.push(row);
+      });
     });
 
-    // total
-    let total_time = final_results.reduce((r, c) => {
-      return r + c.time;
-    }, 0);
+    // compute total
+    finalResultsList.forEach((finalResults) => {
+      // total
+      let total_time = finalResults.reduce((r, c) => {
+        return r + c.time;
+      }, 0);
 
-    table.push({total: total_time});
+      table.push({total: total_time});
 
-    console.log(chalk.white.bold('Benchmark Results:'));
-    console.log(table.toString());
+      console.log(chalk.white.bold('Benchmark Results:'));
+      console.log(table.toString());
+    });
 
-    if (this.opts.send) {
-      if (this.opts.verbose_level > 0) {
-        console.log('Sending data to Keen.IO...');
+    // send all
+    finalResultsList.forEach((finalResults) => {
+      if (this.opts.send) {
+        this._sendData(finalResults);
+      } else {
+        console.error(chalk.green('Benchmark finished. No data was sent.'));
+        return 0;
       }
-      return BB.Promise.mapSeries(final_results, (result) => {
-        // send each result to Keen.IO
-        return this.sendData.send('profiling', result);
-      })
-      .then((results) => {
-        // check if all data was sent to Keen.IO
-        let success_results = results.filter((item) => {
-          return (item.created === true);
-        });
-        if (success_results.length === results.length) {
-          console.log(chalk.green('Benchmark finished. All data sent to Keen.IO.'));
-          return 0;
-        } else {
-          console.log(chalk.red('Benchmark finished. Some data was not sent to Keen.IO:'));
-          console.log(results);
-          return 1;
-        }
-      });
-    } else {
-      console.error(chalk.green('Benchmark finished. No data was sent.'));
-      return 0;
+    });
+  }
+
+  _sendData(finalResults) {
+    if (this.opts.verboseLevel > 0) {
+      console.log('Sending data to Keen.IO...');
     }
+    return BB.Promise.mapSeries(finalResults, (result) => {
+      // send each result to Keen.IO
+      return this.sendData.send('profiling', result);
+    })
+    .then((results) => {
+      // check if all data was sent to Keen.IO
+      let successResults = results.filter((item) => {
+        return (item.created === true);
+      });
+      if (successResults.length === results.length) {
+        console.log(chalk.green('Benchmark finished. All data sent to Keen.IO.'));
+        return 0;
+      } else {
+        console.log(chalk.red('Benchmark finished. Some data was not sent to Keen.IO:'));
+        console.log(results);
+        return 1;
+      }
+    });
   }
 }