+++ src/utils/which_helper.js
@@ -1,12 +1,12 @@
 import BB from 'bluebird';
 import which from 'which';
 
-let whichHelper = (command_path) => {
+let whichHelper = (commandPath) => {
   return new BB.Promise((resolve, reject) => {
-    which(command_path, (err, resolvedPath) => {
+    which(commandPath, (err, resolvedPath) => {
       if (err) {
-        // err is returned if no "command_path" is found on the PATH
-        err = new Error(`'${command_path}' not found in PATH: ${process.env.PATH}. Original error: ${err.message}`);
+        // err is returned if no "commandPath" is found on the PATH
+        err = new Error(`'${commandPath}' not found in PATH: ${process.env.PATH}. Original error: ${err.message}`);
         reject(err);
       } else {
         // if it is found, then the absolute path to the exec is returned